add Repository Interface

+++ App/Http/Requests/UserRequest.php
@@ -16,7 +16,7 @@ class UserRequest
             'email' => trim($_POST['email']) ?? '',
             'gender' => trim($_POST['gender']) ?? '',
             'password' => trim($_POST['password']) ?? '',
-            'role_id' => trim($_POST['role_id']) ?? 1
+            'role_id' => (int) trim($_POST['role_id']) ?? 1
         ];
     }
 
@@ -27,7 +27,7 @@ class UserRequest
             'last_name' => trim($_POST['last_name']),
             'email' => trim($_POST['email']),
             'gender' => trim($_POST['gender']),
-            'role_id' => trim($_POST['role_id'])
+            'role_id' => (int) trim($_POST['role_id'])
         ];
     }
 }
